+++ apps/explorer/src/pages/index.tsx
@@ -5,6 +5,7 @@
 import { wrapCreateBrowserRouter } from '@sentry/react';
 import { createBrowserRouter, Navigate, useLocation, useParams } from 'react-router-dom';
 import { AddressResultPage } from './address-result/AddressResult';
+import { Analytics } from './analytics/Analytics';
 import { AbstractAccountResultPage } from './abstract-account-result/AbstractAccountResult';
 import { CheckpointDetail } from './checkpoints/CheckpointDetail';
 import { CookiePolicyPage } from './cookie-policy/CookiePolicyPage';
@@ -48,6 +49,7 @@ export const router = sentryCreateBrowserRouter([
             { path: 'address/:id', element: <AddressResultPage /> },
             { path: 'account/:id', element: <AbstractAccountResultPage /> },
             { path: 'validators', element: <ValidatorPageResult /> },
+            { path: 'analytics', element: <Analytics /> },
             { path: 'validator/:id', element: <ValidatorDetails /> },
             { path: 'identity/:id', element: <IdentityResult /> },
             { path: 'notarization/:id', element: <NotarizationResult /> },