+++ dapps/sponsored-transactions/src/utils/rpc.ts
@@ -2,6 +2,6 @@
 // Modifications Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { getFullnodeUrl, IotaClient } from '@iota/iota-sdk/client';
+import { getRpcUrl, IotaClient } from '@iota/iota-sdk/client';
 
-export const client = new IotaClient({ url: getFullnodeUrl('testnet') });
+export const client = new IotaClient({ url: getRpcUrl('testnet') });