+++ sdk/create-dapp/templates/react-client-dapp/src/networkConfig.ts
@@ -1,15 +1,15 @@
-import { getFullnodeUrl } from '@iota/iota-sdk/client';
+import { getRpcUrl } from '@iota/iota-sdk/client';
 import { createNetworkConfig } from '@iota/dapp-kit';
 
 const { networkConfig, useNetworkVariable, useNetworkVariables } = createNetworkConfig({
     devnet: {
-        url: getFullnodeUrl('devnet'),
+        url: getRpcUrl('devnet'),
     },
     testnet: {
-        url: getFullnodeUrl('testnet'),
+        url: getRpcUrl('testnet'),
     },
     mainnet: {
-        url: getFullnodeUrl('mainnet'),
+        url: getRpcUrl('mainnet'),
     },
 });